Added folder with the goal of screen builder
