+++ ToDoList/ToDoList/Controllers/TaskTableViewHeaderFooterView/TaskTableViewHeaderFooterView.swift
@@ -1,0 +1,20 @@
+//
+//  TaskTableViewHeaderFooterView.swift
+//  ToDoList
+//
+//  Created by Никита Данилович on 03.05.2023.
+//
+
+import UIKit
+
+class TaskTableViewHeaderFooterView: UITableViewHeaderFooterView {
+
+    /*
+    // Only override draw() if you perform custom drawing.
+    // An empty implementation adversely affects performance during animation.
+    override func draw(_ rect: CGRect) {
+        // Drawing code
+    }
+    */
+
+}
