+++ ToDoList/ToDoList/Xib/TaskManagerViewControllerImp.xib
@@ -8,7 +8,7 @@
         <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
     </dependencies>
     <objects>
-        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="TaskManagerViewController" customModule="ToDoList" customModuleProvider="target">
+        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="TaskManagerViewControllerImp" customModule="ToDoList" customModuleProvider="target">
             <connections>
                 <outlet property="tasksTableView" destination="Z4H-zC-qwr" id="3I3-cp-r4H"/>
                 <outlet property="view" destination="i5M-Pr-FkT" id="BT6-lF-sVU"/>
@@ -19,16 +19,16 @@
             <rect key="frame" x="0.0" y="0.0" width="393" height="852"/>
             <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
             <subviews>
-                <tableView clipsSubviews="YES" contentMode="scaleToFill" alwaysBounceVertical="YES" style="plain" separatorStyle="default" rowHeight="-1" estimatedRowHeight="-1" sectionHeaderHeight="-1" estimatedSectionHeaderHeight="-1" sectionFooterHeight="-1" estimatedSectionFooterHeight="-1" translatesAutoresizingMaskIntoConstraints="NO" id="Z4H-zC-qwr">
-                    <rect key="frame" x="0.0" y="59" width="393" height="759"/>
-                    <color key="backgroundColor" white="0.33333333329999998" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
+                <tableView clipsSubviews="YES" contentMode="scaleAspectFit" alwaysBounceVertical="YES" style="plain" rowHeight="-1" estimatedRowHeight="-1" sectionHeaderHeight="28" estimatedSectionHeaderHeight="-1" sectionFooterHeight="28" estimatedSectionFooterHeight="-1" translatesAutoresizingMaskIntoConstraints="NO" id="Z4H-zC-qwr">
+                    <rect key="frame" x="0.0" y="99" width="393" height="719"/>
+                    <color key="tintColor" red="0.83137254900000002" green="0.30980392159999998" blue="0.30980392159999998" alpha="1" colorSpace="calibratedRGB"/>
                 </tableView>
             </subviews>
             <viewLayoutGuide key="safeArea" id="fnl-2z-Ty3"/>
             <color key="backgroundColor" white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
             <color key="tintColor" red="0.83137254900000002" green="0.30980392159999998" blue="0.30980392159999998" alpha="1" colorSpace="calibratedRGB"/>
             <constraints>
-                <constraint firstItem="Z4H-zC-qwr" firstAttribute="top" secondItem="fnl-2z-Ty3" secondAttribute="top" id="AkB-0U-LIL"/>
+                <constraint firstItem="Z4H-zC-qwr" firstAttribute="top" secondItem="i5M-Pr-FkT" secondAttribute="top" constant="99" id="AkB-0U-LIL"/>
                 <constraint firstItem="Z4H-zC-qwr" firstAttribute="leading" secondItem="fnl-2z-Ty3" secondAttribute="leading" id="M2S-a2-rBO"/>
                 <constraint firstItem="fnl-2z-Ty3" firstAttribute="bottom" secondItem="Z4H-zC-qwr" secondAttribute="bottom" id="fMW-Jw-mrD"/>
                 <constraint firstItem="fnl-2z-Ty3" firstAttribute="trailing" secondItem="Z4H-zC-qwr" secondAttribute="trailing" id="n9H-qg-3Rb"/>
